requirements.txt
openai==1.76.0
pypdf2==3.0.1
python-docx==0.8.11
gradio==4.0.0
pandas==2.0.3
beautifulsoup4==4.12.2
requests==2.31.0
pytest==7.4.3
pytest-cov==4.1.0
mypy==1.6.1
pylint==3.0.2
black==23.9.1
python-dotenv==1.0.0